requirements.txt
absl-py==0.7.1
aiohttp==3.6.2
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
Automat==0.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.13.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
elasticsearch==7.7.1
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-core==4.5.0
jupyter-nbextensions-configurator==0.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
notedown==1.5.1
numpy==1.17.0
opt-einsum==3.1.0
pandas==0.25.1
pandoc-attributes==0.1.7
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyquery==1.4.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.4
queuelib==1.5.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.7.4
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
Twisted==19.7.0
urllib3==1.25.6
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
yarl==1.4.2
zope.interface==4.6.0
pytesseract